[feat] : add complete preview
